+++ packages/components/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @baseapp-frontend/components
 
+## 0.0.19
+
+### Patch Changes
+
+- Add Active Tab functionality to the `messages` module.
+- Tweaks on `MessageItem` styles so it have the correct `max-width`.
+- Updated dependencies
+  - @baseapp-frontend/design-system@0.0.20
+
 ## 0.0.18
 
 ### Patch Changes
@@ -7,7 +16,6 @@
 - Added the CreateChatRoom components for the messages module
 - Updated dependencies
   - @baseapp-frontend/design-system@0.0.19
-
 
 ## 0.0.17
