+++ packages/components/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@baseapp-frontend/components",
   "description": "BaseApp components modules such as comments, notifications, messages, and more.",
-  "version": "0.0.18",
+  "version": "0.0.19",
   "main": "./index.ts",
   "types": "dist/index.d.ts",
   "sideEffects": false,
@@ -21,37 +21,37 @@
     "test:unit": "jest --config ./jest.config.ts"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.6.0",
-    "@mui/lab": "^5.0.0-alpha.170",
-    "@mui/material": "^5.15.19",
-    "@mui/system": "^5.15.19",
-    "@storybook/react": "^8.2.8",
-    "@tanstack/react-query": "^5.45.1",
+    "@hookform/resolvers": "catalog:",
+    "@mui/lab": "catalog:material-ui",
+    "@mui/material": "catalog:material-ui",
+    "@mui/system": "catalog:material-ui",
+    "@storybook/react": "catalog:storybook",
+    "@tanstack/react-query": "catalog:",
     "framer-motion": "^11.5.4",
-    "graphql": "^16.8.1",
-    "js-cookie": "^3.0.5",
-    "luxon": "^3.4.4",
-    "next": "14.3.0-canary.24",
-    "react-hook-form": "^7.51.5",
-    "react-relay": "^16.2.0",
-    "react-virtuoso": "^4.7.11",
-    "relay-runtime": "^16.2.0",
+    "graphql": "catalog:graphql",
+    "js-cookie": "catalog:",
+    "luxon": "catalog:",
+    "next": "catalog:",
+    "react-hook-form": "catalog:",
+    "react-relay": "catalog:graphql",
+    "react-virtuoso": "catalog:",
+    "relay-runtime": "catalog:graphql",
     "use-long-press": "^3.2.0",
-    "zod": "^3.23.8",
-    "zustand": "^4.5.2"
+    "zod": "catalog:",
+    "zustand": "catalog:"
   },
   "peerDependencies": {
     "@baseapp-frontend/authentication": "workspace:*",
     "@baseapp-frontend/design-system": "workspace:*",
     "@baseapp-frontend/graphql": "workspace:*",
     "@baseapp-frontend/utils": "workspace:*",
-    "react": "18.3.1",
-    "react-dom": "18.3.1"
+    "react": "catalog:react18",
+    "react-dom": "catalog:react18"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.24.7",
-    "@babel/preset-react": "^7.24.7",
-    "@babel/preset-typescript": "^7.24.7",
+    "@babel/preset-env": "catalog:storybook",
+    "@babel/preset-react": "catalog:storybook",
+    "@babel/preset-typescript": "catalog:storybook",
     "@baseapp-frontend/config": "workspace:*",
     "@baseapp-frontend/test": "workspace:*",
     "@baseapp-frontend/tsconfig": "workspace:*",
@@ -78,6 +78,7 @@
     "@types/react": "catalog:react18",
     "@types/react-dom": "catalog:react18",
     "@types/react-relay": "catalog:graphql",
+    "@types/relay-runtime": "catalog:graphql",
     "@types/relay-test-utils": "catalog:graphql",
     "autoprefixer": "catalog:tailwind",
     "babel-jest": "catalog:test",
@@ -89,26 +90,26 @@
     "cypress-wait-until": "catalog:test",
     "dotenv": "^16.4.5",
     "dotenv-cli": "^7.4.2",
-    "eslint-plugin-storybook": "^0.8.0",
-    "get-graphql-schema": "^2.1.2",
-    "html-webpack-plugin": "^5.6.0",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "msw": "^2.3.1",
-    "msw-storybook-addon": "^2.0.2",
-    "postcss": "^8.4.41",
-    "postcss-loader": "^8.1.1",
-    "relay-compiler": "^16.2.0",
-    "relay-test-utils": "^17.0.0",
-    "storybook": "^8.2.8",
-    "style-loader": "^4.0.0",
-    "tailwindcss": "^3.4.4",
-    "ts-jest": "^29.1.4",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.4.5",
-    "webpack": "^5.93.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.0.4"
+    "eslint-plugin-storybook": "catalog:lint",
+    "get-graphql-schema": "catalog:graphql",
+    "html-webpack-plugin": "catalog:storybook",
+    "jest": "catalog:test",
+    "jest-environment-jsdom": "catalog:test",
+    "msw": "catalog:storybook",
+    "msw-storybook-addon": "catalog:storybook",
+    "postcss": "catalog:tailwind",
+    "postcss-loader": "catalog:storybook",
+    "relay-compiler": "catalog:graphql",
+    "relay-test-utils": "catalog:graphql",
+    "storybook": "catalog:storybook",
+    "style-loader": "catalog:storybook",
+    "tailwindcss": "catalog:tailwind",
+    "ts-jest": "catalog:test",
+    "ts-node": "catalog:test",
+    "typescript": "catalog:",
+    "webpack": "catalog:storybook",
+    "webpack-cli": "catalog:storybook",
+    "webpack-dev-server": "catalog:storybook"
   },
   "license": "MIT",
   "repository": {
