+++ packages/design-system/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@baseapp-frontend/design-system",
   "description": "Design System components and configurations.",
-  "version": "0.0.19",
+  "version": "0.0.20",
   "main": "./index.ts",
   "types": "dist/index.d.ts",
   "sideEffects": false,
@@ -14,60 +14,60 @@
     "storybook:build": "storybook build"
   },
   "dependencies": {
-    "@emotion/cache": "^11.11.0",
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
+    "@emotion/cache": "catalog:material-ui",
+    "@emotion/react": "catalog:material-ui",
+    "@emotion/styled": "catalog:material-ui",
     "@iconify/react": "^5.0.1",
-    "@mui/icons-material": "^5.15.19",
-    "@mui/lab": "^5.0.0-alpha.170",
-    "@mui/material": "^5.15.19",
-    "@mui/material-nextjs": "^6.1.4",
-    "@mui/system": "^5.15.19",
-    "@mui/x-date-pickers": "^7.6.2",
-    "@storybook/react": "^8.2.8",
+    "@mui/icons-material": "catalog:material-ui",
+    "@mui/lab": "catalog:material-ui",
+    "@mui/material": "catalog:material-ui",
+    "@mui/material-nextjs": "catalog:material-ui",
+    "@mui/system": "catalog:material-ui",
+    "@mui/x-date-pickers": "catalog:material-ui",
+    "@storybook/react": "catalog:storybook",
     "framer-motion": "^11.2.10",
-    "jotai": "^2.9.3",
-    "lodash": "^4.17.21",
-    "next": "14.3.0-canary.24",
+    "jotai": "catalog:",
+    "lodash": "catalog:",
+    "next": "catalog:",
     "react-dropzone": "^14.2.3",
-    "react-hook-form": "^7.51.5",
+    "react-hook-form": "catalog:",
     "react-lazy-load-image-component": "^1.6.2",
     "simplebar-react": "^3.2.5"
   },
   "peerDependencies": {
     "@baseapp-frontend/utils": "workspace:*",
-    "react": "18.3.1",
-    "react-dom": "18.3.1"
+    "react": "catalog:react18",
+    "react-dom": "catalog:react18"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.24.7",
-    "@babel/preset-react": "^7.24.7",
-    "@babel/preset-typescript": "^7.24.7",
+    "@babel/preset-env": "catalog:storybook",
+    "@babel/preset-react": "catalog:storybook",
+    "@babel/preset-typescript": "catalog:storybook",
     "@baseapp-frontend/config": "workspace:*",
     "@baseapp-frontend/tsconfig": "workspace:*",
-    "@chromatic-com/storybook": "^1.5.0",
-    "@storybook/addon-docs": "^8.2.8",
-    "@storybook/addon-essentials": "^8.2.8",
-    "@storybook/addon-interactions": "^8.2.8",
-    "@storybook/addon-links": "^8.2.8",
-    "@storybook/addon-styling-webpack": "^1.0.0",
-    "@storybook/addon-webpack5-compiler-babel": "^3.0.3",
-    "@storybook/blocks": "^8.2.8",
-    "@storybook/react-webpack5": "^8.2.8",
-    "@storybook/test": "^8.2.8",
-    "@tailwindcss/typography": "^0.5.13",
-    "@types/lodash": "^4.17.7",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "autoprefixer": "^10.4.19",
-    "css-loader": "^7.1.2",
-    "eslint-plugin-storybook": "^0.8.0",
-    "postcss": "^8.4.41",
-    "postcss-loader": "^8.1.1",
-    "storybook": "^8.2.8",
-    "style-loader": "^4.0.0",
-    "tailwindcss": "^3.4.4",
-    "typescript": "^5.4.5"
+    "@chromatic-com/storybook": "catalog:storybook",
+    "@storybook/addon-docs": "catalog:storybook",
+    "@storybook/addon-essentials": "catalog:storybook",
+    "@storybook/addon-interactions": "catalog:storybook",
+    "@storybook/addon-links": "catalog:storybook",
+    "@storybook/addon-styling-webpack": "catalog:storybook",
+    "@storybook/addon-webpack5-compiler-babel": "catalog:storybook",
+    "@storybook/blocks": "catalog:storybook",
+    "@storybook/react-webpack5": "catalog:storybook",
+    "@storybook/test": "catalog:storybook",
+    "@tailwindcss/typography": "catalog:tailwind",
+    "@types/lodash": "catalog:",
+    "@types/react": "catalog:react18",
+    "@types/react-dom": "catalog:react18",
+    "autoprefixer": "catalog:tailwind",
+    "css-loader": "catalog:storybook",
+    "eslint-plugin-storybook": "catalog:lint",
+    "postcss": "catalog:tailwind",
+    "postcss-loader": "catalog:storybook",
+    "storybook": "catalog:storybook",
+    "style-loader": "catalog:storybook",
+    "tailwindcss": "catalog:tailwind",
+    "typescript": "catalog:"
   },
   "license": "MIT",
   "repository": {
